wasm: Split up wasm main package over several files
Split up wasm main package over several files, as pkg/wasm/main.go was
getting a little big. In a follow up commit we will refactor
evaluator.Runtime into an interface and implement it in this package as
well. It will go with the imports.go file. This commit is a mechanical
preparation for it moving source code and improving on some comments.
